feat : move files in app > _components to src > components folder
p.s auth 트리에 종속된 CheckBox 컴포넌트 파일도 함께 옮겼습니다.
