--- apps/client/src/app/_components/Form/index.ts
@@ -1,2 +1,0 @@
-export { FormElement } from "@components/Form/FormElement";
-export type { FormProps } from "@components/Form/FormProps.type";
