+++ apps/client/src/app/auth/_components/Form/index.ts
@@ -1,2 +1,1 @@
 export { BusinessApiButton } from "@auth/_components/Form/BusinessApiButton";
-export { CheckBox } from "@auth/_components/Form/CheckBox";
