+++ apps/client/src/components/CheckBox/index.ts
@@ -1,0 +1,1 @@
+export * from "./CheckBox";
