+++ apps/client/src/components/Coupon/CouponSummary/CouponSummary.constants.ts
