+++ apps/client/src/components/Coupon/ExpansionCoupon/ExpansionCoupon.stories.tsx
