+++ apps/client/src/components/Coupon/ExpansionCoupon/ExpansionCoupon.types.ts
