+++ apps/client/src/components/Coupon/ReceiveCoupon/ReceiveCoupon.stories.tsx
