+++ apps/client/src/components/Coupon/ReceiveCoupon/ReceiveCoupon.types.ts
