+++ apps/client/src/components/Coupon/components/CouponComponents.types.ts
