+++ apps/client/src/components/Coupon/index.ts
@@ -1,4 +1,4 @@
-export * from "./ClaimCoupon";
 export * from "./Coupon";
 export * from "./CouponSummary";
 export * from "./ExpansionCoupon";
+export * from "./ReceiveCoupon";
