+++ apps/client/src/components/StoreItem/StoreItem.stories.tsx
