+++ apps/client/src/components/StoreItem/StoreItem.types.ts
