+++ apps/client/src/components/index.ts
@@ -1,4 +1,5 @@
+export * from "./CheckBox";
 export * from "./Coupon";
+export * from "./Form";
 export * from "./HorizontalLine";
 export * from "./StoreItem";
-export * from "./Form";
