+++ apps/client/src/mocks/Apis/my-page/types.ts
@@ -1,5 +1,5 @@
-import type { StoreProps } from "~/app/_components";
 import type { CouponSummaryProps } from "~/app/my-page/_components";
+import type { StoreProps } from "~/components";
 
 export interface MyPageBuyerProps {
   nearbyStores: StoreProps[];
